Fixed byte alignment issue in endgame screen

+++ Chess.Engine/src/Chess.Engine.Core/src/InputSource/WinUIInputSource.cpp
@@ -76,7 +76,7 @@ void WinUIInputSource::onGameEnded(EndGameState state, Side winner)
 {
 	EndgameStateEvent event;
 	event.state	 = state;
-	event.winner = winner;
+	event.winner = static_cast<int>(winner);
 
 	sendToUI(MessageType::EndGameState, &event);
 }
